Feature: Adding upload files.

+++ app/Controllers/HunterController.php
@@ -3,12 +3,13 @@
 namespace App\Controllers;
 
 use App\Controllers\BaseController;
-use App\Models\{HunterModel,RecompensadoModel,TipoHunterModel,TipoSanguineoModel,TipoNenModel};
+use App\Models\{AvatarHunterModel,HunterModel,RecompensadoModel,TipoHunterModel,TipoSanguineoModel,TipoNenModel};
 use App\Validation\HunterValidation;
 use CILogViewer\CILogViewer;
 
 class HunterController extends BaseController
 {
+    
     public function index()
     {
         $model = new HunterModel();
@@ -90,10 +91,22 @@ class HunterController extends BaseController
             'inicio' => $this->request->getVar('inicio'),
             'termino' => $this->request->getVar('termino'),
         ];
-        $model->insert($dados_tratados);
+        $hunter_id = $model->insert($dados_tratados);
         $nome_inteiro = $dados_tratados['nome_hunter'];
         $primeiro_nome = explode(' ', $dados_tratados['nome_hunter'])[0];
         log_message('notice', "Hunter $nome_inteiro foi adicionado.");
+        if ($this->request->getFiles()) {
+            $avatar_hunter_model = new AvatarHunterModel();
+            foreach ($this->request->getFiles()['avatar'] as $avatar) {
+                $nome_original = $avatar->getName();
+                $caminho_imagem = 'uploads/avatars/' . $hunter_id . '/' . $nome_original;
+                $avatar->move(WRITEPATH . 'uploads/avatars/' . $hunter_id, $nome_original);
+                $avatar_hunter_model->insert([
+                    'hunter_id' => $hunter_id,
+                    'imagem' => $caminho_imagem
+                ]);
+            }
+        }
         return redirect()->to(route_to('indexHunter'))->with('success', "Hunter $primeiro_nome foi inserido com sucesso.");
     }
 
@@ -166,6 +179,19 @@ class HunterController extends BaseController
             'termino' => $this->request->getVar('termino'),
         ];
         $model->update($id, $dados_tratados);
+        if ($this->request->getFileMultiple('avatar')[0]->isValid()) {
+            $avatar_hunter_model = new AvatarHunterModel();
+            $avatar_hunter_model->where('hunter_id', $id)->delete();
+            foreach ($this->request->getFileMultiple('avatar') as $avatar) {
+                $nome_original = $avatar->getName();
+                $caminho_imagem = 'uploads/avatars/' . $id . '/' . $nome_original;
+                $avatar->move(WRITEPATH . 'uploads/avatars/' . $id, $nome_original);
+                $avatar_hunter_model->insert([
+                    'hunter_id' => $id,
+                    'imagem' => $caminho_imagem,
+                ]);
+            }
+        }
         $primeiro_nome = explode(' ', $dados_tratados['nome_hunter'])[0];
         $nome_inteiro = $dados_tratados['nome_hunter'];
         log_message('info', "Hunter $nome_inteiro obteve atualizações.");
@@ -177,9 +203,22 @@ class HunterController extends BaseController
         $model = new HunterModel();
         $registro = $model->find($id);
         $primeiro_nome = explode(' ', $registro['nome_hunter'])[0];
-        $model->delete($id);
-        log_message('alert', "Hunter $primeiro_nome foi enviado(a) para a lixeira.");
-        return redirect()->to(route_to('indexHunter'))->with('warning', "Hunter $primeiro_nome foi enviado(a) para a lixeira.");
+        if ($registro) {
+            $caminho_imagem = WRITEPATH . 'uploads/avatars/' . $id;
+            $trash_avatars_caminho = WRITEPATH . 'uploads/trash/avatars/' . $id;
+            if (!is_dir($trash_avatars_caminho)) {
+                mkdir($trash_avatars_caminho, 0777, true);
+            }
+            $this->moveDirectory($caminho_imagem, $trash_avatars_caminho);
+            if (is_dir($caminho_imagem)) {
+                rmdir($caminho_imagem);
+            }
+            $model->delete($id);
+            log_message('alert', "Hunter $primeiro_nome foi enviado(a) para a lixeira.");
+            return redirect()->to(route_to('indexHunter'))->with('warning', "Hunter $primeiro_nome foi enviado(a) para a lixeira.");
+        } else {
+            return redirect()->to(route_to('indexHunter'))->with('error', "Registro de caçador não encontrado.");
+        }
     }
 
     public function logsView()
@@ -236,6 +275,14 @@ class HunterController extends BaseController
         $nome_hunter = $registro_deletado['nome_hunter'];
         if ($registro_deletado) {
             $model->onlyDeleted()->builder()->update(['deleted_at' => null], ['id' => $id]);
+            $caminho_imagem = WRITEPATH . 'uploads/avatars/' . $id;
+            $trash_avatars_caminho = WRITEPATH . 'uploads/trash/avatars/' . $id;
+            if (is_dir($trash_avatars_caminho)) {
+                rename($trash_avatars_caminho, $caminho_imagem);
+                if (is_dir($trash_avatars_caminho)) {
+                    rmdir($trash_avatars_caminho);
+                }
+            }
             log_message('info', "Hunter $nome_hunter foi restaurado(a).");
             return redirect()->to(route_to('trashHunter'))->with('success', "Hunter $nome_hunter retornou para a listagem principal.");
         } else {
@@ -254,8 +301,49 @@ class HunterController extends BaseController
             return redirect()->to(route_to('trashHunter'))->with('warning', "Não é possível excluir $nome_hunter permanentemente, pois está associado em $quantidade_hunters registro(s) de recompensados.");
         }
         $model->onlyDeleted()->where('id', $id)->purgeDeleted();
+        $trash_avatars_caminho = WRITEPATH . 'uploads/trash/avatars/' . $id;
+        $this->deleteDirectoryContents($trash_avatars_caminho);
+        if (is_dir($trash_avatars_caminho)) {
+            rmdir($trash_avatars_caminho);
+        }
         log_message('alert', "Hunter $nome_hunter foi excluído(a) permanentemente.");
         return redirect()->to(route_to('trashHunter'))->with('danger', "Hunter $nome_hunter foi excluído(a) permanentemente.");
+    }
+
+    private function moveDirectory($source, $destination) {
+        if (!is_dir($source)) {
+            return false;
+        }
+        if (!is_dir($destination)) {
+            mkdir($destination, 0777, true);
+        }
+        $directory = dir($source);
+        while (false !== ($entry = $directory->read())) {
+            if ($entry == '.' || $entry == '..') {
+                continue;
+            }
+            $sourcePath = $source . '/' . $entry;
+            $destinationPath = $destination . '/' . $entry;
+            if (is_dir($sourcePath)) {
+                $this->moveDirectory($sourcePath, $destinationPath);
+                rmdir($sourcePath);
+            } else {
+                rename($sourcePath, $destinationPath);
+            }
+        }
+        $directory->close();
+        return true;
+    }
+
+    private function deleteDirectoryContents($dir) {
+        foreach(glob($dir . '/*') as $file) {
+            if(is_dir($file)) { 
+                $this->deleteDirectoryContents($file);
+                rmdir($file);
+            } else {
+                unlink($file);
+            }
+        }
     }
 
 }
